Hoàn thành luồng kho confirm báo giá

+++ src/main/java/fpt/edu/vn/gms/repository/PurchaseRequestItemRepository.java
@@ -3,5 +3,5 @@ package fpt.edu.vn.gms.repository;
 import fpt.edu.vn.gms.entity.PurchaseRequestItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRequestItemRepository extends JpaRepository<PurchaseRequestItem, Integer> {
+public interface PurchaseRequestItemRepository extends JpaRepository<PurchaseRequestItem, Long> {
 }
