+++ src/main/java/fpt/edu/vn/gms/repository/PurchaseRequestRepository.java
@@ -3,5 +3,5 @@ package fpt.edu.vn.gms.repository;
 import fpt.edu.vn.gms.entity.PurchaseRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRequestRepository extends JpaRepository<PurchaseRequest, Integer> {
+public interface PurchaseRequestRepository extends JpaRepository<PurchaseRequest, Long> {
 }
